+++ lib/libesp32/berry_matter/src/be_matter_module.c
@@ -118,12 +118,20 @@ const char* matter_get_opcode_name(uint16_t opcode) {
 BE_FUNC_CTYPE_DECLARE(matter_get_opcode_name, "s", "i")
 
 const char* matter_get_attribute_name(uint16_t cluster, uint16_t attribute) {
-  for (const matter_cluster_t * cl = matterAllClusters; cl->id != 0xFFFF; cl++) {
-    if (cl->id == cluster) {
-      for (const matter_attribute_t * at = cl->attributes; at->id != 0xFFFF; at++) {
-        if (at->id == attribute) {
-          return at->name;
+  if (attribute < 0xFFF0) {
+    for (const matter_cluster_t * cl = matterAllClusters; cl->id != 0xFFFF; cl++) {
+      if (cl->id == cluster) {
+        for (const matter_attribute_t * at = cl->attributes; at->id != 0xFFFF; at++) {
+          if (at->id == attribute) {
+            return at->name;
+          }
         }
+      }
+    }
+  } else {
+    for (const matter_attribute_t * at = matter_Attributes_Common; at->id != 0xFFFF; at++) {
+      if (at->id == attribute) {
+        return at->name;
       }
     }
   }
@@ -132,11 +140,13 @@ const char* matter_get_attribute_name(uint16_t cluster, uint16_t attribute) {
 BE_FUNC_CTYPE_DECLARE(matter_get_attribute_name, "s", "ii")
 
 bbool matter_is_attribute_writable(uint16_t cluster, uint16_t attribute) {
-  for (const matter_cluster_t * cl = matterAllClusters; cl->id != 0xFFFF; cl++) {
-    if (cl->id == cluster) {
-      for (const matter_attribute_t * at = cl->attributes; at->id != 0xFFFF; at++) {
-        if (at->id == attribute) {
-          return (at->flags & 0x01) ? btrue : bfalse;
+  if (attribute < 0xFFF0) {
+    for (const matter_cluster_t * cl = matterAllClusters; cl->id != 0xFFFF; cl++) {
+      if (cl->id == cluster) {
+        for (const matter_attribute_t * at = cl->attributes; at->id != 0xFFFF; at++) {
+          if (at->id == attribute) {
+            return (at->flags & 0x01) ? btrue : bfalse;
+          }
         }
       }
     }
@@ -146,14 +156,18 @@ bbool matter_is_attribute_writable(uint16_t cluster, uint16_t attribute) {
 BE_FUNC_CTYPE_DECLARE(matter_is_attribute_writable, "b", "ii")
 
 bbool matter_is_attribute_reportable(uint16_t cluster, uint16_t attribute) {
-  for (const matter_cluster_t * cl = matterAllClusters; cl->id != 0xFFFF; cl++) {
-    if (cl->id == cluster) {
-      for (const matter_attribute_t * at = cl->attributes; at->id != 0xFFFF; at++) {
-        if (at->id == attribute) {
-          return (at->flags & 0x02) ? btrue : bfalse;
+  if (attribute < 0xFFF0) {
+    for (const matter_cluster_t * cl = matterAllClusters; cl->id != 0xFFFF; cl++) {
+      if (cl->id == cluster) {
+        for (const matter_attribute_t * at = cl->attributes; at->id != 0xFFFF; at++) {
+          if (at->id == attribute) {
+            return (at->flags & 0x02) ? btrue : bfalse;
+          }
         }
       }
     }
+  } else {
+    return btrue;
   }
   return bfalse;
 }