+++ lib/libesp32/berry_matter/src/solidify/solidified_Matter_Plugin_1_Device.h
@@ -19,7 +19,7 @@ be_local_closure(Matter_Plugin_Device_read_attribute,   /* name */
     0,                          /* has sup protos */
     NULL,                       /* no sub protos */
     1,                          /* has constants */
-    ( &(const bvalue[37]) {     /* constants */
+    ( &(const bvalue[36]) {     /* constants */
     /* K0   */  be_nested_str_weak(matter),
     /* K1   */  be_nested_str_weak(TLV),
     /* K2   */  be_nested_str_weak(cluster),
@@ -30,43 +30,42 @@ be_local_closure(Matter_Plugin_Device_read_attribute,   /* name */
     /* K7   */  be_nested_str_weak(U2),
     /* K8   */  be_const_int(1),
     /* K9   */  be_nested_str_weak(U1),
-    /* K10  */  be_nested_str_weak(U4),
-    /* K11  */  be_nested_str_weak(Matter_TLV_array),
-    /* K12  */  be_nested_str_weak(TYPES),
-    /* K13  */  be_nested_str_weak(keys),
-    /* K14  */  be_nested_str_weak(add_struct),
-    /* K15  */  be_nested_str_weak(add_TLV),
-    /* K16  */  be_nested_str_weak(stop_iteration),
-    /* K17  */  be_nested_str_weak(NON_BRIDGE_VENDOR),
-    /* K18  */  be_nested_str_weak(find),
-    /* K19  */  be_nested_str_weak(get_admin_vendor),
-    /* K20  */  be_nested_str_weak(device),
-    /* K21  */  be_nested_str_weak(disable_bridge_mode),
-    /* K22  */  be_nested_str_weak(read_attribute),
-    /* K23  */  be_nested_str_weak(string),
-    /* K24  */  be_nested_str_weak(UTF1),
-    /* K25  */  be_nested_str_weak(tasmota),
-    /* K26  */  be_nested_str_weak(cmd),
-    /* K27  */  be_nested_str_weak(DeviceName),
-    /* K28  */  be_nested_str_weak(get_name),
-    /* K29  */  be_nested_str_weak(Status_X202),
-    /* K30  */  be_nested_str_weak(StatusFWR),
-    /* K31  */  be_nested_str_weak(Version),
-    /* K32  */  be_nested_str_weak(_X28),
-    /* K33  */  be_nested_str_weak(wifi),
-    /* K34  */  be_nested_str_weak(mac),
-    /* K35  */  be_nested_str_weak(),
-    /* K36  */  be_nested_str_weak(BOOL),
+    /* K10  */  be_nested_str_weak(Matter_TLV_array),
+    /* K11  */  be_nested_str_weak(TYPES),
+    /* K12  */  be_nested_str_weak(keys),
+    /* K13  */  be_nested_str_weak(add_struct),
+    /* K14  */  be_nested_str_weak(add_TLV),
+    /* K15  */  be_nested_str_weak(stop_iteration),
+    /* K16  */  be_nested_str_weak(NON_BRIDGE_VENDOR),
+    /* K17  */  be_nested_str_weak(find),
+    /* K18  */  be_nested_str_weak(get_admin_vendor),
+    /* K19  */  be_nested_str_weak(device),
+    /* K20  */  be_nested_str_weak(disable_bridge_mode),
+    /* K21  */  be_nested_str_weak(string),
+    /* K22  */  be_nested_str_weak(UTF1),
+    /* K23  */  be_nested_str_weak(tasmota),
+    /* K24  */  be_nested_str_weak(cmd),
+    /* K25  */  be_nested_str_weak(DeviceName),
+    /* K26  */  be_nested_str_weak(get_name),
+    /* K27  */  be_nested_str_weak(Status_X202),
+    /* K28  */  be_nested_str_weak(StatusFWR),
+    /* K29  */  be_nested_str_weak(Version),
+    /* K30  */  be_nested_str_weak(_X28),
+    /* K31  */  be_nested_str_weak(wifi),
+    /* K32  */  be_nested_str_weak(mac),
+    /* K33  */  be_nested_str_weak(),
+    /* K34  */  be_nested_str_weak(BOOL),
+    /* K35  */  be_nested_str_weak(read_attribute),
     }),
     be_str_weak(read_attribute),
     &be_const_str_solidified,
-    ( &(const binstruction[254]) {  /* code */
+    ( &(const binstruction[189]) {  /* code */
       0xB8120000,  //  0000  GETNGBL	R4	K0
       0x88100901,  //  0001  GETMBR	R4	R4	K1
       0x88140502,  //  0002  GETMBR	R5	R2	K2
       0x88180503,  //  0003  GETMBR	R6	R2	K3
       0x1C1C0B04,  //  0004  EQ	R7	R5	K4
-      0x781E0021,  //  0005  JMPF	R7	#0028
+      0x781E000F,  //  0005  JMPF	R7	#0016
       0x1C1C0D05,  //  0006  EQ	R7	R6	K5
       0x781E0005,  //  0007  JMPF	R7	#000E
       0x8C1C0706,  //  0008  GETMET	R7	R3	K6
@@ -74,247 +73,182 @@ be_local_closure(Matter_Plugin_Device_read_attribute,   /* name */
       0x58280005,  //  000A  LDCONST	R10	K5
       0x7C1C0600,  //  000B  CALL	R7	3
       0x80040E00,  //  000C  RET	1	R7
-      0x70020018,  //  000D  JMP		#0027
+      0x70020006,  //  000D  JMP		#0015
       0x1C1C0D08,  //  000E  EQ	R7	R6	K8
-      0x781E0005,  //  000F  JMPF	R7	#0016
+      0x781E0004,  //  000F  JMPF	R7	#0015
       0x8C1C0706,  //  0010  GETMET	R7	R3	K6
       0x88240909,  //  0011  GETMBR	R9	R4	K9
       0x58280005,  //  0012  LDCONST	R10	K5
       0x7C1C0600,  //  0013  CALL	R7	3
       0x80040E00,  //  0014  RET	1	R7
-      0x70020010,  //  0015  JMP		#0027
-      0x541EFFFB,  //  0016  LDINT	R7	65532
-      0x1C1C0C07,  //  0017  EQ	R7	R6	R7
-      0x781E0005,  //  0018  JMPF	R7	#001F
-      0x8C1C0706,  //  0019  GETMET	R7	R3	K6
-      0x8824090A,  //  001A  GETMBR	R9	R4	K10
-      0x58280005,  //  001B  LDCONST	R10	K5
-      0x7C1C0600,  //  001C  CALL	R7	3
-      0x80040E00,  //  001D  RET	1	R7
-      0x70020007,  //  001E  JMP		#0027
-      0x541EFFFC,  //  001F  LDINT	R7	65533
-      0x1C1C0C07,  //  0020  EQ	R7	R6	R7
-      0x781E0004,  //  0021  JMPF	R7	#0027
-      0x8C1C0706,  //  0022  GETMET	R7	R3	K6
-      0x8824090A,  //  0023  GETMBR	R9	R4	K10
-      0x542A0003,  //  0024  LDINT	R10	4
-      0x7C1C0600,  //  0025  CALL	R7	3
-      0x80040E00,  //  0026  RET	1	R7
-      0x700200D4,  //  0027  JMP		#00FD
-      0x541E0003,  //  0028  LDINT	R7	4
-      0x1C1C0A07,  //  0029  EQ	R7	R5	R7
-      0x781E0016,  //  002A  JMPF	R7	#0042
-      0x1C1C0D05,  //  002B  EQ	R7	R6	K5
-      0x781E0002,  //  002C  JMPF	R7	#0030
-      0x4C1C0000,  //  002D  LDNIL	R7
-      0x80040E00,  //  002E  RET	1	R7
-      0x70020010,  //  002F  JMP		#0041
-      0x541EFFFB,  //  0030  LDINT	R7	65532
-      0x1C1C0C07,  //  0031  EQ	R7	R6	R7
-      0x781E0005,  //  0032  JMPF	R7	#0039
-      0x8C1C0706,  //  0033  GETMET	R7	R3	K6
-      0x8824090A,  //  0034  GETMBR	R9	R4	K10
-      0x58280005,  //  0035  LDCONST	R10	K5
-      0x7C1C0600,  //  0036  CALL	R7	3
-      0x80040E00,  //  0037  RET	1	R7
-      0x70020007,  //  0038  JMP		#0041
-      0x541EFFFC,  //  0039  LDINT	R7	65533
-      0x1C1C0C07,  //  003A  EQ	R7	R6	R7
-      0x781E0004,  //  003B  JMPF	R7	#0041
-      0x8C1C0706,  //  003C  GETMET	R7	R3	K6
-      0x8824090A,  //  003D  GETMBR	R9	R4	K10
-      0x542A0003,  //  003E  LDINT	R10	4
-      0x7C1C0600,  //  003F  CALL	R7	3
-      0x80040E00,  //  0040  RET	1	R7
-      0x700200BA,  //  0041  JMP		#00FD
-      0x541E0004,  //  0042  LDINT	R7	5
-      0x1C1C0A07,  //  0043  EQ	R7	R5	R7
-      0x781E0011,  //  0044  JMPF	R7	#0057
-      0x541EFFFB,  //  0045  LDINT	R7	65532
-      0x1C1C0C07,  //  0046  EQ	R7	R6	R7
-      0x781E0005,  //  0047  JMPF	R7	#004E
-      0x8C1C0706,  //  0048  GETMET	R7	R3	K6
-      0x8824090A,  //  0049  GETMBR	R9	R4	K10
-      0x58280005,  //  004A  LDCONST	R10	K5
-      0x7C1C0600,  //  004B  CALL	R7	3
-      0x80040E00,  //  004C  RET	1	R7
-      0x70020007,  //  004D  JMP		#0056
-      0x541EFFFC,  //  004E  LDINT	R7	65533
-      0x1C1C0C07,  //  004F  EQ	R7	R6	R7
-      0x781E0004,  //  0050  JMPF	R7	#0056
-      0x8C1C0706,  //  0051  GETMET	R7	R3	K6
-      0x8824090A,  //  0052  GETMBR	R9	R4	K10
-      0x542A0003,  //  0053  LDINT	R10	4
-      0x7C1C0600,  //  0054  CALL	R7	3
-      0x80040E00,  //  0055  RET	1	R7
-      0x700200A5,  //  0056  JMP		#00FD
-      0x541E001C,  //  0057  LDINT	R7	29
-      0x1C1C0A07,  //  0058  EQ	R7	R5	R7
-      0x781E003E,  //  0059  JMPF	R7	#0099
-      0x1C1C0D05,  //  005A  EQ	R7	R6	K5
-      0x781E0032,  //  005B  JMPF	R7	#008F
-      0x8C1C090B,  //  005C  GETMET	R7	R4	K11
-      0x7C1C0200,  //  005D  CALL	R7	1
-      0x8820010C,  //  005E  GETMBR	R8	R0	K12
-      0x60240010,  //  005F  GETGBL	R9	G16
-      0x8C28110D,  //  0060  GETMET	R10	R8	K13
-      0x7C280200,  //  0061  CALL	R10	1
-      0x7C240200,  //  0062  CALL	R9	1
-      0xA802000E,  //  0063  EXBLK	0	#0073
-      0x5C281200,  //  0064  MOVE	R10	R9
-      0x7C280000,  //  0065  CALL	R10	0
-      0x8C2C0F0E,  //  0066  GETMET	R11	R7	K14
-      0x7C2C0200,  //  0067  CALL	R11	1
-      0x8C30170F,  //  0068  GETMET	R12	R11	K15
-      0x58380005,  //  0069  LDCONST	R14	K5
-      0x883C0907,  //  006A  GETMBR	R15	R4	K7
-      0x5C401400,  //  006B  MOVE	R16	R10
-      0x7C300800,  //  006C  CALL	R12	4
-      0x8C30170F,  //  006D  GETMET	R12	R11	K15
-      0x58380008,  //  006E  LDCONST	R14	K8
-      0x883C0907,  //  006F  GETMBR	R15	R4	K7
-      0x9440100A,  //  0070  GETIDX	R16	R8	R10
-      0x7C300800,  //  0071  CALL	R12	4
-      0x7001FFF0,  //  0072  JMP		#0064
-      0x58240010,  //  0073  LDCONST	R9	K16
-      0xAC240200,  //  0074  CATCH	R9	1	0
-      0xB0080000,  //  0075  RAISE	2	R0	R0
-      0x88240111,  //  0076  GETMBR	R9	R0	K17
-      0x8C241312,  //  0077  GETMET	R9	R9	K18
-      0x8C2C0313,  //  0078  GETMET	R11	R1	K19
-      0x7C2C0200,  //  0079  CALL	R11	1
-      0x7C240400,  //  007A  CALL	R9	2
-      0x4C280000,  //  007B  LDNIL	R10
-      0x1C24120A,  //  007C  EQ	R9	R9	R10
-      0x7826000E,  //  007D  JMPF	R9	#008D
-      0x88240114,  //  007E  GETMBR	R9	R0	K20
-      0x88241315,  //  007F  GETMBR	R9	R9	K21
-      0x7426000B,  //  0080  JMPT	R9	#008D
-      0x8C240F0E,  //  0081  GETMET	R9	R7	K14
-      0x7C240200,  //  0082  CALL	R9	1
-      0x8C28130F,  //  0083  GETMET	R10	R9	K15
-      0x58300005,  //  0084  LDCONST	R12	K5
-      0x88340907,  //  0085  GETMBR	R13	R4	K7
-      0x543A0012,  //  0086  LDINT	R14	19
-      0x7C280800,  //  0087  CALL	R10	4
-      0x8C28130F,  //  0088  GETMET	R10	R9	K15
-      0x58300008,  //  0089  LDCONST	R12	K8
-      0x88340907,  //  008A  GETMBR	R13	R4	K7
-      0x58380008,  //  008B  LDCONST	R14	K8
-      0x7C280800,  //  008C  CALL	R10	4
-      0x80040E00,  //  008D  RET	1	R7
-      0x70020008,  //  008E  JMP		#0098
-      0x601C0003,  //  008F  GETGBL	R7	G3
-      0x5C200000,  //  0090  MOVE	R8	R0
-      0x7C1C0200,  //  0091  CALL	R7	1
-      0x8C1C0F16,  //  0092  GETMET	R7	R7	K22
-      0x5C240200,  //  0093  MOVE	R9	R1
-      0x5C280400,  //  0094  MOVE	R10	R2
-      0x5C2C0600,  //  0095  MOVE	R11	R3
-      0x7C1C0800,  //  0096  CALL	R7	4
-      0x80040E00,  //  0097  RET	1	R7
-      0x70020063,  //  0098  JMP		#00FD
-      0x541E0038,  //  0099  LDINT	R7	57
-      0x1C1C0A07,  //  009A  EQ	R7	R5	R7
-      0x781E0057,  //  009B  JMPF	R7	#00F4
-      0xA41E2E00,  //  009C  IMPORT	R7	K23
-      0x1C200D04,  //  009D  EQ	R8	R6	K4
-      0x7822000A,  //  009E  JMPF	R8	#00AA
-      0x8C200706,  //  009F  GETMET	R8	R3	K6
-      0x88280918,  //  00A0  GETMBR	R10	R4	K24
-      0xB82E3200,  //  00A1  GETNGBL	R11	K25
-      0x8C2C171A,  //  00A2  GETMET	R11	R11	K26
-      0x5834001B,  //  00A3  LDCONST	R13	K27
-      0x50380200,  //  00A4  LDBOOL	R14	1	0
-      0x7C2C0600,  //  00A5  CALL	R11	3
-      0x942C171B,  //  00A6  GETIDX	R11	R11	K27
-      0x7C200600,  //  00A7  CALL	R8	3
-      0x80041000,  //  00A8  RET	1	R8
-      0x70020048,  //  00A9  JMP		#00F3
-      0x54220004,  //  00AA  LDINT	R8	5
-      0x1C200C08,  //  00AB  EQ	R8	R6	R8
-      0x78220006,  //  00AC  JMPF	R8	#00B4
-      0x8C200706,  //  00AD  GETMET	R8	R3	K6
-      0x88280918,  //  00AE  GETMBR	R10	R4	K24
-      0x8C2C011C,  //  00AF  GETMET	R11	R0	K28
-      0x7C2C0200,  //  00B0  CALL	R11	1
-      0x7C200600,  //  00B1  CALL	R8	3
-      0x80041000,  //  00B2  RET	1	R8
-      0x7002003E,  //  00B3  JMP		#00F3
-      0x54220009,  //  00B4  LDINT	R8	10
-      0x1C200C08,  //  00B5  EQ	R8	R6	R8
-      0x78220015,  //  00B6  JMPF	R8	#00CD
-      0xB8223200,  //  00B7  GETNGBL	R8	K25
-      0x8C20111A,  //  00B8  GETMET	R8	R8	K26
-      0x5828001D,  //  00B9  LDCONST	R10	K29
-      0x502C0200,  //  00BA  LDBOOL	R11	1	0
-      0x7C200600,  //  00BB  CALL	R8	3
-      0x9420111E,  //  00BC  GETIDX	R8	R8	K30
-      0x9420111F,  //  00BD  GETIDX	R8	R8	K31
-      0x8C240F12,  //  00BE  GETMET	R9	R7	K18
-      0x5C2C1000,  //  00BF  MOVE	R11	R8
-      0x58300020,  //  00C0  LDCONST	R12	K32
-      0x7C240600,  //  00C1  CALL	R9	3
-      0x24281305,  //  00C2  GT	R10	R9	K5
-      0x782A0002,  //  00C3  JMPF	R10	#00C7
-      0x04281308,  //  00C4  SUB	R10	R9	K8
-      0x402A0A0A,  //  00C5  CONNECT	R10	K5	R10
-      0x9420100A,  //  00C6  GETIDX	R8	R8	R10
-      0x8C280706,  //  00C7  GETMET	R10	R3	K6
-      0x88300918,  //  00C8  GETMBR	R12	R4	K24
-      0x5C341000,  //  00C9  MOVE	R13	R8
-      0x7C280600,  //  00CA  CALL	R10	3
-      0x80041400,  //  00CB  RET	1	R10
-      0x70020025,  //  00CC  JMP		#00F3
-      0x5422000E,  //  00CD  LDINT	R8	15
-      0x1C200C08,  //  00CE  EQ	R8	R6	R8
-      0x7822000B,  //  00CF  JMPF	R8	#00DC
-      0x8C200706,  //  00D0  GETMET	R8	R3	K6
-      0x88280918,  //  00D1  GETMBR	R10	R4	K24
-      0xB82E3200,  //  00D2  GETNGBL	R11	K25
-      0x8C2C1721,  //  00D3  GETMET	R11	R11	K33
-      0x7C2C0200,  //  00D4  CALL	R11	1
-      0x8C2C1712,  //  00D5  GETMET	R11	R11	K18
-      0x58340022,  //  00D6  LDCONST	R13	K34
-      0x58380023,  //  00D7  LDCONST	R14	K35
-      0x7C2C0600,  //  00D8  CALL	R11	3
-      0x7C200600,  //  00D9  CALL	R8	3
-      0x80041000,  //  00DA  RET	1	R8
-      0x70020016,  //  00DB  JMP		#00F3
-      0x54220010,  //  00DC  LDINT	R8	17
-      0x1C200C08,  //  00DD  EQ	R8	R6	R8
-      0x78220005,  //  00DE  JMPF	R8	#00E5
-      0x8C200706,  //  00DF  GETMET	R8	R3	K6
-      0x88280924,  //  00E0  GETMBR	R10	R4	K36
-      0x582C0008,  //  00E1  LDCONST	R11	K8
-      0x7C200600,  //  00E2  CALL	R8	3
-      0x80041000,  //  00E3  RET	1	R8
-      0x7002000D,  //  00E4  JMP		#00F3
-      0x54220011,  //  00E5  LDINT	R8	18
-      0x1C200C08,  //  00E6  EQ	R8	R6	R8
-      0x7822000A,  //  00E7  JMPF	R8	#00F3
-      0x8C200706,  //  00E8  GETMET	R8	R3	K6
-      0x88280918,  //  00E9  GETMBR	R10	R4	K24
-      0xB82E3200,  //  00EA  GETNGBL	R11	K25
-      0x8C2C1721,  //  00EB  GETMET	R11	R11	K33
-      0x7C2C0200,  //  00EC  CALL	R11	1
-      0x8C2C1712,  //  00ED  GETMET	R11	R11	K18
-      0x58340022,  //  00EE  LDCONST	R13	K34
-      0x58380023,  //  00EF  LDCONST	R14	K35
-      0x7C2C0600,  //  00F0  CALL	R11	3
-      0x7C200600,  //  00F1  CALL	R8	3
-      0x80041000,  //  00F2  RET	1	R8
-      0x70020008,  //  00F3  JMP		#00FD
-      0x601C0003,  //  00F4  GETGBL	R7	G3
-      0x5C200000,  //  00F5  MOVE	R8	R0
-      0x7C1C0200,  //  00F6  CALL	R7	1
-      0x8C1C0F16,  //  00F7  GETMET	R7	R7	K22
-      0x5C240200,  //  00F8  MOVE	R9	R1
-      0x5C280400,  //  00F9  MOVE	R10	R2
-      0x5C2C0600,  //  00FA  MOVE	R11	R3
-      0x7C1C0800,  //  00FB  CALL	R7	4
-      0x80040E00,  //  00FC  RET	1	R7
-      0x80000000,  //  00FD  RET	0
+      0x7002009D,  //  0015  JMP		#00B4
+      0x541E0003,  //  0016  LDINT	R7	4
+      0x1C1C0A07,  //  0017  EQ	R7	R5	R7
+      0x781E0004,  //  0018  JMPF	R7	#001E
+      0x1C1C0D05,  //  0019  EQ	R7	R6	K5
+      0x781E0001,  //  001A  JMPF	R7	#001D
+      0x4C1C0000,  //  001B  LDNIL	R7
+      0x80040E00,  //  001C  RET	1	R7
+      0x70020095,  //  001D  JMP		#00B4
+      0x541E0004,  //  001E  LDINT	R7	5
+      0x1C1C0A07,  //  001F  EQ	R7	R5	R7
+      0x781E0000,  //  0020  JMPF	R7	#0022
+      0x70020091,  //  0021  JMP		#00B4
+      0x541E001C,  //  0022  LDINT	R7	29
+      0x1C1C0A07,  //  0023  EQ	R7	R5	R7
+      0x781E0034,  //  0024  JMPF	R7	#005A
+      0x1C1C0D05,  //  0025  EQ	R7	R6	K5
+      0x781E0031,  //  0026  JMPF	R7	#0059
+      0x8C1C090A,  //  0027  GETMET	R7	R4	K10
+      0x7C1C0200,  //  0028  CALL	R7	1
+      0x8820010B,  //  0029  GETMBR	R8	R0	K11
+      0x60240010,  //  002A  GETGBL	R9	G16
+      0x8C28110C,  //  002B  GETMET	R10	R8	K12
+      0x7C280200,  //  002C  CALL	R10	1
+      0x7C240200,  //  002D  CALL	R9	1
+      0xA802000E,  //  002E  EXBLK	0	#003E
+      0x5C281200,  //  002F  MOVE	R10	R9
+      0x7C280000,  //  0030  CALL	R10	0
+      0x8C2C0F0D,  //  0031  GETMET	R11	R7	K13
+      0x7C2C0200,  //  0032  CALL	R11	1
+      0x8C30170E,  //  0033  GETMET	R12	R11	K14
+      0x58380005,  //  0034  LDCONST	R14	K5
+      0x883C0907,  //  0035  GETMBR	R15	R4	K7
+      0x5C401400,  //  0036  MOVE	R16	R10
+      0x7C300800,  //  0037  CALL	R12	4
+      0x8C30170E,  //  0038  GETMET	R12	R11	K14
+      0x58380008,  //  0039  LDCONST	R14	K8
+      0x883C0907,  //  003A  GETMBR	R15	R4	K7
+      0x9440100A,  //  003B  GETIDX	R16	R8	R10
+      0x7C300800,  //  003C  CALL	R12	4
+      0x7001FFF0,  //  003D  JMP		#002F
+      0x5824000F,  //  003E  LDCONST	R9	K15
+      0xAC240200,  //  003F  CATCH	R9	1	0
+      0xB0080000,  //  0040  RAISE	2	R0	R0
+      0x88240110,  //  0041  GETMBR	R9	R0	K16
+      0x8C241311,  //  0042  GETMET	R9	R9	K17
+      0x8C2C0312,  //  0043  GETMET	R11	R1	K18
+      0x7C2C0200,  //  0044  CALL	R11	1
+      0x7C240400,  //  0045  CALL	R9	2
+      0x4C280000,  //  0046  LDNIL	R10
+      0x1C24120A,  //  0047  EQ	R9	R9	R10
+      0x7826000E,  //  0048  JMPF	R9	#0058
+      0x88240113,  //  0049  GETMBR	R9	R0	K19
+      0x88241314,  //  004A  GETMBR	R9	R9	K20
+      0x7426000B,  //  004B  JMPT	R9	#0058
+      0x8C240F0D,  //  004C  GETMET	R9	R7	K13
+      0x7C240200,  //  004D  CALL	R9	1
+      0x8C28130E,  //  004E  GETMET	R10	R9	K14
+      0x58300005,  //  004F  LDCONST	R12	K5
+      0x88340907,  //  0050  GETMBR	R13	R4	K7
+      0x543A0012,  //  0051  LDINT	R14	19
+      0x7C280800,  //  0052  CALL	R10	4
+      0x8C28130E,  //  0053  GETMET	R10	R9	K14
+      0x58300008,  //  0054  LDCONST	R12	K8
+      0x88340907,  //  0055  GETMBR	R13	R4	K7
+      0x58380008,  //  0056  LDCONST	R14	K8
+      0x7C280800,  //  0057  CALL	R10	4
+      0x80040E00,  //  0058  RET	1	R7
+      0x70020059,  //  0059  JMP		#00B4
+      0x541E0038,  //  005A  LDINT	R7	57
+      0x1C1C0A07,  //  005B  EQ	R7	R5	R7
+      0x781E0056,  //  005C  JMPF	R7	#00B4
+      0xA41E2A00,  //  005D  IMPORT	R7	K21
+      0x1C200D04,  //  005E  EQ	R8	R6	K4
+      0x7822000A,  //  005F  JMPF	R8	#006B
+      0x8C200706,  //  0060  GETMET	R8	R3	K6
+      0x88280916,  //  0061  GETMBR	R10	R4	K22
+      0xB82E2E00,  //  0062  GETNGBL	R11	K23
+      0x8C2C1718,  //  0063  GETMET	R11	R11	K24
+      0x58340019,  //  0064  LDCONST	R13	K25
+      0x50380200,  //  0065  LDBOOL	R14	1	0
+      0x7C2C0600,  //  0066  CALL	R11	3
+      0x942C1719,  //  0067  GETIDX	R11	R11	K25
+      0x7C200600,  //  0068  CALL	R8	3
+      0x80041000,  //  0069  RET	1	R8
+      0x70020048,  //  006A  JMP		#00B4
+      0x54220004,  //  006B  LDINT	R8	5
+      0x1C200C08,  //  006C  EQ	R8	R6	R8
+      0x78220006,  //  006D  JMPF	R8	#0075
+      0x8C200706,  //  006E  GETMET	R8	R3	K6
+      0x88280916,  //  006F  GETMBR	R10	R4	K22
+      0x8C2C011A,  //  0070  GETMET	R11	R0	K26
+      0x7C2C0200,  //  0071  CALL	R11	1
+      0x7C200600,  //  0072  CALL	R8	3
+      0x80041000,  //  0073  RET	1	R8
+      0x7002003E,  //  0074  JMP		#00B4
+      0x54220009,  //  0075  LDINT	R8	10
+      0x1C200C08,  //  0076  EQ	R8	R6	R8
+      0x78220015,  //  0077  JMPF	R8	#008E
+      0xB8222E00,  //  0078  GETNGBL	R8	K23
+      0x8C201118,  //  0079  GETMET	R8	R8	K24
+      0x5828001B,  //  007A  LDCONST	R10	K27
+      0x502C0200,  //  007B  LDBOOL	R11	1	0
+      0x7C200600,  //  007C  CALL	R8	3
+      0x9420111C,  //  007D  GETIDX	R8	R8	K28
+      0x9420111D,  //  007E  GETIDX	R8	R8	K29
+      0x8C240F11,  //  007F  GETMET	R9	R7	K17
+      0x5C2C1000,  //  0080  MOVE	R11	R8
+      0x5830001E,  //  0081  LDCONST	R12	K30
+      0x7C240600,  //  0082  CALL	R9	3
+      0x24281305,  //  0083  GT	R10	R9	K5
+      0x782A0002,  //  0084  JMPF	R10	#0088
+      0x04281308,  //  0085  SUB	R10	R9	K8
+      0x402A0A0A,  //  0086  CONNECT	R10	K5	R10
+      0x9420100A,  //  0087  GETIDX	R8	R8	R10
+      0x8C280706,  //  0088  GETMET	R10	R3	K6
+      0x88300916,  //  0089  GETMBR	R12	R4	K22
+      0x5C341000,  //  008A  MOVE	R13	R8
+      0x7C280600,  //  008B  CALL	R10	3
+      0x80041400,  //  008C  RET	1	R10
+      0x70020025,  //  008D  JMP		#00B4
+      0x5422000E,  //  008E  LDINT	R8	15
+      0x1C200C08,  //  008F  EQ	R8	R6	R8
+      0x7822000B,  //  0090  JMPF	R8	#009D
+      0x8C200706,  //  0091  GETMET	R8	R3	K6
+      0x88280916,  //  0092  GETMBR	R10	R4	K22
+      0xB82E2E00,  //  0093  GETNGBL	R11	K23
+      0x8C2C171F,  //  0094  GETMET	R11	R11	K31
+      0x7C2C0200,  //  0095  CALL	R11	1
+      0x8C2C1711,  //  0096  GETMET	R11	R11	K17
+      0x58340020,  //  0097  LDCONST	R13	K32
+      0x58380021,  //  0098  LDCONST	R14	K33
+      0x7C2C0600,  //  0099  CALL	R11	3
+      0x7C200600,  //  009A  CALL	R8	3
+      0x80041000,  //  009B  RET	1	R8
+      0x70020016,  //  009C  JMP		#00B4
+      0x54220010,  //  009D  LDINT	R8	17
+      0x1C200C08,  //  009E  EQ	R8	R6	R8
+      0x78220005,  //  009F  JMPF	R8	#00A6
+      0x8C200706,  //  00A0  GETMET	R8	R3	K6
+      0x88280922,  //  00A1  GETMBR	R10	R4	K34
+      0x582C0008,  //  00A2  LDCONST	R11	K8
+      0x7C200600,  //  00A3  CALL	R8	3
+      0x80041000,  //  00A4  RET	1	R8
+      0x7002000D,  //  00A5  JMP		#00B4
+      0x54220011,  //  00A6  LDINT	R8	18
+      0x1C200C08,  //  00A7  EQ	R8	R6	R8
+      0x7822000A,  //  00A8  JMPF	R8	#00B4
+      0x8C200706,  //  00A9  GETMET	R8	R3	K6
+      0x88280916,  //  00AA  GETMBR	R10	R4	K22
+      0xB82E2E00,  //  00AB  GETNGBL	R11	K23
+      0x8C2C171F,  //  00AC  GETMET	R11	R11	K31
+      0x7C2C0200,  //  00AD  CALL	R11	1
+      0x8C2C1711,  //  00AE  GETMET	R11	R11	K17
+      0x58340020,  //  00AF  LDCONST	R13	K32
+      0x58380021,  //  00B0  LDCONST	R14	K33
+      0x7C2C0600,  //  00B1  CALL	R11	3
+      0x7C200600,  //  00B2  CALL	R8	3
+      0x80041000,  //  00B3  RET	1	R8
+      0x601C0003,  //  00B4  GETGBL	R7	G3
+      0x5C200000,  //  00B5  MOVE	R8	R0
+      0x7C1C0200,  //  00B6  CALL	R7	1
+      0x8C1C0F23,  //  00B7  GETMET	R7	R7	K35
+      0x5C240200,  //  00B8  MOVE	R9	R1
+      0x5C280400,  //  00B9  MOVE	R10	R2
+      0x5C2C0600,  //  00BA  MOVE	R11	R3
+      0x7C1C0800,  //  00BB  CALL	R7	4
+      0x80040E00,  //  00BC  RET	1	R7
     })
   )
 );
@@ -610,7 +544,7 @@ be_local_class(Matter_Plugin_Device,
         be_const_map( *     be_nested_map(5,
     ( (struct bmapnode*) &(const bmapnode[]) {
         { be_const_key_int(5, -1), be_const_simple_instance(be_nested_simple_instance(&be_class_list, {
-        be_const_list( *     be_nested_list(8,
+        be_const_list( *     be_nested_list(12,
     ( (struct bvalue*) &(const bvalue[]) {
         be_const_int(0),
         be_const_int(1),
@@ -618,21 +552,29 @@ be_local_class(Matter_Plugin_Device,
         be_const_int(3),
         be_const_int(4),
         be_const_int(5),
+        be_const_int(65528),
+        be_const_int(65529),
+        be_const_int(65530),
+        be_const_int(65531),
         be_const_int(65532),
         be_const_int(65533),
     }))    ) } )) },
         { be_const_key_int(29, -1), be_const_simple_instance(be_nested_simple_instance(&be_class_list, {
-        be_const_list( *     be_nested_list(6,
+        be_const_list( *     be_nested_list(10,
     ( (struct bvalue*) &(const bvalue[]) {
         be_const_int(0),
         be_const_int(1),
         be_const_int(2),
         be_const_int(3),
+        be_const_int(65528),
+        be_const_int(65529),
+        be_const_int(65530),
+        be_const_int(65531),
         be_const_int(65532),
         be_const_int(65533),
     }))    ) } )) },
         { be_const_key_int(57, -1), be_const_simple_instance(be_nested_simple_instance(&be_class_list, {
-        be_const_list( *     be_nested_list(6,
+        be_const_list( *     be_nested_list(12,
     ( (struct bvalue*) &(const bvalue[]) {
         be_const_int(3),
         be_const_int(5),
@@ -640,19 +582,33 @@ be_local_class(Matter_Plugin_Device,
         be_const_int(15),
         be_const_int(17),
         be_const_int(18),
+        be_const_int(65528),
+        be_const_int(65529),
+        be_const_int(65530),
+        be_const_int(65531),
+        be_const_int(65532),
+        be_const_int(65533),
     }))    ) } )) },
         { be_const_key_int(3, -1), be_const_simple_instance(be_nested_simple_instance(&be_class_list, {
-        be_const_list( *     be_nested_list(4,
+        be_const_list( *     be_nested_list(8,
     ( (struct bvalue*) &(const bvalue[]) {
         be_const_int(0),
         be_const_int(1),
+        be_const_int(65528),
+        be_const_int(65529),
+        be_const_int(65530),
+        be_const_int(65531),
         be_const_int(65532),
         be_const_int(65533),
     }))    ) } )) },
         { be_const_key_int(4, 1), be_const_simple_instance(be_nested_simple_instance(&be_class_list, {
-        be_const_list( *     be_nested_list(3,
+        be_const_list( *     be_nested_list(7,
     ( (struct bvalue*) &(const bvalue[]) {
         be_const_int(0),
+        be_const_int(65528),
+        be_const_int(65529),
+        be_const_int(65530),
+        be_const_int(65531),
         be_const_int(65532),
         be_const_int(65533),
     }))    ) } )) },